Function lonlat_bng::drop_float_array

pub extern fn drop_float_array(lons: Array, lats: Array)

Free memory which Rust has allocated across the FFI boundary (f64 values)

Examples

let lon_vec: Vec<f64> = vec![-2.0183041005533306];
let lat_vec: Vec<f64> = vec![54.589097162646141];
let lon_arr = Array {
    data: lon_vec.as_ptr() as *const libc::c_void,
    len: lon_vec.len() as libc::size_t,
};
let lat_arr = Array {
    data: lat_vec.as_ptr() as *const libc::c_void,
    len: lat_vec.len() as libc::size_t,
};
let (eastings, northings) = convert_to_bng_threaded(lon_arr, lat_arr);
drop_float_array(eastings, northings);

An example FFI implementation is available at Convertbng, specifically in the _void_array_to_list function.

Safety

This function is unsafe because it accesses a raw pointer which could contain arbitrary data